+++ lib/constant.ts
@@ -1,0 +1,4 @@
+export const EaseConstant = {
+  Slash: '/',
+  Dot: '.',
+};